Redo Vector2 Cover tests for TinyGen
